Improvement to Custom Actions. Only deleted custom actions related to the site hierarchy.
Improvement to Custom Actions. Only deleted custom actions related to
the site hierarchy. This way you can install stuff from different site
hierarchy files in the same site collection.

+++ Sherpa.Library/SiteHierarchy/CustomActionsManager.cs
@@ -11,7 +11,7 @@ namespace Sherpa.Library.SiteHierarchy
     {
         private static readonly ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-        public void SetUpCustomActions(ClientContext context, List<ShCustomAction> customActions)
+        public void SetUpCustomActions(ClientContext context, string CustomActionsPrefix, List<ShCustomAction> customActions)
         {
             Log.Info("Adding custom actions");
             Site site = context.Site;
@@ -22,7 +22,10 @@ namespace Sherpa.Library.SiteHierarchy
             for (var i = site.UserCustomActions.Count - 1; i >= 0; i--)
             {
                 var customAction = site.UserCustomActions[i];
-                customAction.DeleteObject();
+                if (customAction.Name.StartsWith(CustomActionsPrefix))
+                {
+                    customAction.DeleteObject();
+                }
             }
 
             if (context.HasPendingRequest)
@@ -45,15 +48,13 @@ namespace Sherpa.Library.SiteHierarchy
                 userCustomAction.Sequence = customAction.Sequence;
                 userCustomAction.ScriptSrc = customAction.ScriptSrc;
                 userCustomAction.ScriptBlock = customAction.ScriptBlock;
-
+                userCustomAction.Name = CustomActionsPrefix + "_" + customAction.ScriptSrc.Split('/')[customAction.ScriptSrc.Split('/').Length - 1].Replace(".", "");
                 userCustomAction.Description = customAction.Description;
-                userCustomAction.RegistrationId = customAction.RegistrationId;
                 userCustomAction.RegistrationType = customAction.RegistrationType;
                 userCustomAction.Title = customAction.Title;
-                userCustomAction.Name = customAction.Name;
                 userCustomAction.ImageUrl = customAction.ImageUrl;
                 userCustomAction.Group = customAction.Group;
-
+                
                 try
                 {
                     userCustomAction.Update();
